Time: 3 ms (6.38%) | Memory: 21.7 MB (94.97%) - LeetSync

+++ 98-validate-binary-search-tree/validate-binary-search-tree.cpp
@@ -14,9 +14,7 @@ public:
     bool checkBST(TreeNode* root, long minVal, long maxVal){
         if(!root) return true;
         if(root->val <= minVal || root->val >= maxVal) return false;
-
-        return (checkBST(root->left, minVal, static_cast<long>(root->val)) && 
-                checkBST(root->right, static_cast<long>(root->val), maxVal));
+        return checkBST(root->left, minVal, root->val) && checkBST(root->right, root->val, maxVal);
     }
     bool isValidBST(TreeNode* root) {
         return checkBST(root, LONG_MIN, LONG_MAX);
